feat(#103): 수익관리 페이지 디자인 변경
