+++ src/app/my/income/_components/AccountInfo.tsx
@@ -1,29 +1,59 @@
-import { BigArrowSVG } from '@/icons/svg';
+import { BigArrowSVG, MoneySVG } from '@/icons/svg';
 
 const data = {
   bank: '우리은행',
   accountNumber: '1002-123-456789',
 };
 
-const AccountInfo = ({ view }: { view: 'main' | 'payment' }) => {
+interface AccountInfoProps {
+  view: 'main' | 'payment';
+  handleApply: () => void;
+}
+const AccountInfo = ({ view, handleApply }: AccountInfoProps) => {
   return view === 'main' ? (
-    <aside className="mt-14 flex h-fit w-full max-w-xs flex-col rounded-[0.31rem] shadow-float">
-      <h2 className="flex h-12 w-full items-center justify-between border-b border-solid border-sub-color4 pl-[1.19rem] text-lg font-semibold text-sub-color3">
-        계좌 정보
-        <BigArrowSVG width="34" height="34" className="fill-sub-color2" />
-      </h2>
-      <ul className="flex h-20 w-full flex-col justify-center gap-[0.81rem] px-[1.19rem] text-sm text-sub-color3">
-        <li className="flex gap-4">
-          <p className="w-14 font-semibold">은행</p> <span>{data.bank}</span>
-        </li>
-        <li className="flex gap-4">
-          <p className="w-14 font-semibold">계좌번호</p>
-          <span>{data.accountNumber}</span>
-        </li>
-      </ul>
+    <aside className="flex max-h-[16.5rem] flex-col gap-3">
+      <section className="flex h-fit w-full max-w-xs flex-col rounded-admin bg-white shadow-float shadow-float">
+        <h2 className="flex h-12 w-full items-center justify-between border-b border-solid border-sub-color4 pl-[1.19rem] text-lg font-semibold text-sub-color3">
+          계좌 정보
+          <BigArrowSVG width="34" height="34" className="fill-sub-color2" />
+        </h2>
+        <ul className="flex h-20 w-full flex-col justify-center gap-[0.81rem] px-[1.19rem] text-sm text-sub-color3">
+          <li className="flex gap-4">
+            <p className="w-14 font-semibold">은행</p> <span>{data.bank}</span>
+          </li>
+          <li className="flex gap-4 whitespace-nowrap">
+            <p className="w-14 font-semibold">계좌번호</p>
+            <span>{data.accountNumber}</span>
+          </li>
+        </ul>
+      </section>
+
+      <section className="flex flex-col rounded-admin bg-white px-5 py-4 shadow-float shadow-float">
+        <ul className="mb-3 flex w-full flex-col gap-1 whitespace-nowrap text-sm text-sub-color3">
+          <li className="flex items-center">
+            <p className="w-10 font-semibold">총 수익</p>
+            <span className="ml-[0.69rem] text-lg font-bold text-black">
+              1,230,800원
+            </span>
+          </li>
+          <li className="flex items-center ">
+            <p className="w-10 font-semibold">이번 달</p>
+            <span className="ml-[0.69rem] text-lg font-bold text-sub-color1">
+              1,230,800원
+            </span>
+          </li>
+        </ul>
+        <button
+          onClick={handleApply}
+          className="flex h-7 w-full items-center justify-center rounded-[0.31rem] bg-main-color text-sm font-semibold text-white"
+        >
+          <MoneySVG width="18" height="18" fill="white" stroke="white" />
+          정산신청
+        </button>
+      </section>
     </aside>
   ) : (
-    <aside className="mt-14 flex h-fit w-full max-w-xs flex-col rounded-[0.31rem] p-4 shadow-float">
+    <aside className="flex h-fit w-full max-w-xs flex-col rounded-admin bg-white p-4 shadow-float">
       <h2 className="mb-2 text-sm font-semibold text-main-color">
         *정산금 지급 기준
       </h2>
