+++ src/app/my/income/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 import { useState } from 'react';
-import { ArrowUpSVG } from '@/icons/svg';
 import AccountInfo from './_components/AccountInfo';
 import IncomeTable from './_components/IncomeDataViewer';
 import IncomeOverview from './_components/IncomeOverview';
@@ -18,25 +17,16 @@ const IncomePage = () => {
   };
 
   return (
-    <div className="mx-auto flex w-max gap-4">
-      <section className="flex w-full max-w-[40rem] flex-col gap-[0.63rem]">
-        <h1 className="mt-3 flex items-center text-2xl font-bold text-sub-color3">
-          {view === 'payment' && (
-            <button onClick={handlePrev} className="origin-center -rotate-90">
-              <ArrowUpSVG fill="black" />
-            </button>
-          )}
-          수익관리
-        </h1>
-
-        <div className="flex flex-col gap-[0.63rem]">
-          <IncomeOverview view={view} handleApply={handleApply} />
+    <>
+      <section className="w-full max-w-[40rem]">
+        <div className="flex flex-col gap-4">
+          <IncomeOverview view={view} handlePrev={handlePrev} />
           {view === 'main' ? <IncomeTable /> : <Payment />}
         </div>
       </section>
 
-      <AccountInfo view={view} />
-    </div>
+      <AccountInfo view={view} handleApply={handleApply} />
+    </>
   );
 };
